Moved CreateCov over to submit_batch
